Move link validation to 11ty after build event. Api documentation failing
